+++ CHANGES.md
@@ -3,6 +3,8 @@ Changes
 
 0.1.1 (unreleased)
 
+- Add --untabify <n> option to magik-lint
+
 
 0.1.0 (2019-06-19)